[AD] Rename some utils ahead of refactoring
