+++ pkg/autodiscovery/common/utils/pod_annotations_test.go
@@ -1,0 +1,367 @@
+package utils
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/DataDog/datadog-agent/pkg/autodiscovery/integration"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestExtractCheckNamesFromPodAnnotations(t *testing.T) {
+	tests := []struct {
+		name         string
+		annotations  map[string]string
+		adIdentifier string
+		checkNames   []string
+	}{
+		{
+			name:         "no check metadata",
+			annotations:  map[string]string{},
+			adIdentifier: "redis",
+			checkNames:   []string{},
+		},
+		{
+			name: "legacy annotations",
+			annotations: map[string]string{
+				"service-discovery.datadoghq.com/redis.check_names": `["redisdb"]`,
+			},
+			adIdentifier: "redis",
+			checkNames:   []string{"redisdb"},
+		},
+		{
+			name: "v1 annotations",
+			annotations: map[string]string{
+				"ad.datadoghq.com/redis.check_names":                `["redisdb"]`,
+				"service-discovery.datadoghq.com/redis.check_names": `["foo"]`,
+			},
+			adIdentifier: "redis",
+			checkNames:   []string{"redisdb"},
+		},
+		{
+			name: "v2 annotations",
+			annotations: map[string]string{
+				"ad.datadoghq.com/redis.checks": `{
+					"redisdb": {}
+				}`,
+				"service-discovery.datadoghq.com/redis.check_names": `["foo"]`,
+				"ad.datadoghq.com/redis.check_names":                `["bar"]`,
+			},
+			adIdentifier: "redis",
+			checkNames:   []string{"redisdb"},
+		},
+		{
+			name: "v2 annotations, multiple checks",
+			annotations: map[string]string{
+				"ad.datadoghq.com/redis.checks": `{
+					"redisdb": {},
+					"foobar": {}
+				}`,
+			},
+			adIdentifier: "redis",
+			checkNames:   []string{"redisdb", "foobar"},
+		},
+	}
+
+	for _, tt := range tests {
+		checkNames, err := ExtractCheckNamesFromPodAnnotations(tt.annotations, tt.adIdentifier)
+		assert.Nil(t, err)
+		assert.ElementsMatch(t, tt.checkNames, checkNames, "check names do not match")
+
+	}
+}
+
+func TestExtractTemplatesFromPodAnnotations(t *testing.T) {
+	const adID = "docker://foobar"
+
+	tests := []struct {
+		name         string
+		annotations  map[string]string
+		adIdentifier string
+		output       []integration.Config
+		errs         []error
+	}{
+		{
+			name: "Nominal case with two templates",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  "[\"apache\",\"http_check\"]",
+				"ad.datadoghq.com/foobar.init_configs": "[{},{}]",
+				"ad.datadoghq.com/foobar.instances":    "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"},{\"name\":\"My service\",\"timeout\":1,\"url\":\"http://%%host%%\"}]",
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data("{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}")},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+				{
+					Name:          "http_check",
+					Instances:     []integration.Data{integration.Data("{\"name\":\"My service\",\"timeout\":1,\"url\":\"http://%%host%%\"}")},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Take one, ignore one",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  "[\"apache\"]",
+				"ad.datadoghq.com/foobar.init_configs": "[{}]",
+				"ad.datadoghq.com/foobar.instances":    "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+				"quux.check_names":                     "[\"apache\"]",
+				"quux.init_configs":                    "[{}]",
+				"quux.instances":                       "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data("{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}")},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Logs config",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.logs": "[{\"service\":\"any_service\",\"source\":\"any_source\"}]",
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					LogsConfig:    integration.Data("[{\"service\":\"any_service\",\"source\":\"any_source\"}]"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Check + logs",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  "[\"apache\"]",
+				"ad.datadoghq.com/foobar.init_configs": "[{}]",
+				"ad.datadoghq.com/foobar.instances":    "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+				"ad.datadoghq.com/foobar.logs":         "[{\"service\":\"any_service\",\"source\":\"any_source\"}]",
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data("{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}")},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+				{
+					LogsConfig:    integration.Data("[{\"service\":\"any_service\",\"source\":\"any_source\"}]"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Missing check_names, silently ignore map",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.init_configs": "[{}]",
+				"ad.datadoghq.com/foobar.instances":    "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+			},
+			adIdentifier: "foobar",
+			output:       nil,
+		},
+		{
+			name: "Missing init_configs, error out",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names": "[\"apache\"]",
+				"ad.datadoghq.com/foobar.instances":   "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+			},
+			adIdentifier: "foobar",
+			output:       nil,
+			errs:         []error{errors.New("could not extract checks config: missing init_configs key")},
+		},
+		{
+			name: "Invalid instances json",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  "[\"apache\"]",
+				"ad.datadoghq.com/foobar.init_configs": "[{}]",
+				"ad.datadoghq.com/foobar.instances":    "[{\"apache_status_url\" \"http://%%host%%/server-status?auto\"}]",
+			},
+			adIdentifier: "foobar",
+			output:       nil,
+			errs:         []error{errors.New("could not extract checks config: in instances: failed to unmarshal JSON: invalid character '\"' after object key")},
+		},
+		{
+			name: "Invalid logs json",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.logs": "{\"service\":\"any_service\",\"source\":\"any_source\"}",
+			},
+			adIdentifier: "foobar",
+			output:       nil,
+			errs:         []error{errors.New("could not extract logs config: invalid format, expected an array, got: 'map[service:any_service source:any_source]'")},
+		},
+		{
+			name: "Invalid checks but valid logs",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names": "[\"apache\"]",
+				"ad.datadoghq.com/foobar.instances":   "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+				"ad.datadoghq.com/foobar.logs":        "[{\"service\":\"any_service\",\"source\":\"any_source\"}]",
+			},
+			adIdentifier: "foobar",
+			errs:         []error{errors.New("could not extract checks config: missing init_configs key")},
+			output: []integration.Config{
+				{
+					LogsConfig:    integration.Data("[{\"service\":\"any_service\",\"source\":\"any_source\"}]"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Invalid checks and invalid logs",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names": "[\"apache\"]",
+				"ad.datadoghq.com/foobar.instances":   "[{\"apache_status_url\":\"http://%%host%%/server-status?auto\"}]",
+				"ad.datadoghq.com/foobar.logs":        "{\"service\":\"any_service\",\"source\":\"any_source\"}",
+			},
+			adIdentifier: "foobar",
+			errs: []error{
+				errors.New("could not extract checks config: missing init_configs key"),
+				errors.New("could not extract logs config: invalid format, expected an array, got: 'map[service:any_service source:any_source]'"),
+			},
+			output: nil,
+		},
+		{
+			name: "Two ways, same config (1)",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  `["apache","apache"]`,
+				"ad.datadoghq.com/foobar.init_configs": `[{},{}]`,
+				"ad.datadoghq.com/foobar.instances":    `[{"apache_status_url":"http://%%host%%/server-status?auto1"},{"apache_status_url":"http://%%host%%/server-status?auto2"}]`,
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto1"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto2"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "Two ways, same config (2)",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.check_names":  `["apache"]`,
+				"ad.datadoghq.com/foobar.init_configs": `[{}]`,
+				"ad.datadoghq.com/foobar.instances":    `[[{"apache_status_url":"http://%%host%%/server-status?auto1"},{"apache_status_url":"http://%%host%%/server-status?auto2"}]]`,
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto1"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto2"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "v2 annotations",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.checks": `{
+					"apache": {
+						"instances": [
+							{"apache_status_url":"http://%%host%%/server-status?auto2"}
+						]
+					}
+				}`,
+				"service-discovery.datadoghq.com/foobar.check_names": `["foo"]`,
+				"ad.datadoghq.com/foobar.check_names":                `["bar"]`,
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto2"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+		{
+			name: "v2 annotations + logs",
+			annotations: map[string]string{
+				"ad.datadoghq.com/foobar.checks": `{
+					"apache": {
+						"instances": [
+							{"apache_status_url":"http://%%host%%/server-status?auto2"}
+						]
+					}
+				}`,
+				"ad.datadoghq.com/foobar.logs": "[{\"service\":\"any_service\",\"source\":\"any_source\"}]",
+			},
+			adIdentifier: "foobar",
+			output: []integration.Config{
+				{
+					Name:          "apache",
+					Instances:     []integration.Data{integration.Data(`{"apache_status_url":"http://%%host%%/server-status?auto2"}`)},
+					InitConfig:    integration.Data("{}"),
+					ADIdentifiers: []string{adID},
+				},
+				{
+					LogsConfig:    integration.Data("[{\"service\":\"any_service\",\"source\":\"any_source\"}]"),
+					ADIdentifiers: []string{adID},
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			configs, errs := ExtractTemplatesFromPodAnnotations(adID, tt.annotations, tt.adIdentifier)
+			assert.ElementsMatch(t, tt.output, configs)
+			assert.ElementsMatch(t, tt.errs, errs)
+		})
+	}
+}
+
+func TestExtractCheckIDFromPodAnnotations(t *testing.T) {
+	tests := []struct {
+		name          string
+		annotations   map[string]string
+		containerName string
+		want          string
+		found         bool
+	}{
+		{
+			name:          "found",
+			annotations:   map[string]string{"ad.datadoghq.com/foo.check.id": "bar"},
+			containerName: "foo",
+			want:          "bar",
+			found:         true,
+		},
+		{
+			name:          "not found",
+			annotations:   map[string]string{"ad.datadoghq.com/foo.check.id": "bar"},
+			containerName: "baz",
+			want:          "",
+			found:         false,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, found := ExtractCheckIDFromPodAnnotations(tt.annotations, tt.containerName)
+			assert.Equal(t, tt.want, got)
+			assert.Equal(t, tt.found, found)
+		})
+	}
+}
